camiyaqui-master/requirements.txt
boto3==1.9.202
botocore==1.12.202
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-coreapi==2.3.0
django-extensions==2.2.1
django-translated-fields==0.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
docutils==0.15.2
exponent-server-sdk==0.3.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
